Split plot types, data-cleaning, and histogram logic into submodules
Carve cohesive chunks out of three oversized plot files and re-export them
from the originals so every consumer import stays unchanged:
- types.ts -> types/{fills,reference-lines,plot-3d}.ts (barrel re-export)
- data-cleaning.ts -> data-cleaning-signal.ts (signal primitives; export *)
- Histogram.svelte -> histogram.ts (pure binning/range helpers, unit-testable)

Dedupe the histogram binning config via shared BinConfig/RangeLimit types and
import the cleaning config types from their canonical data-cleaning-signal home.
