+++ src/lib/plot/core/types.ts
@@ -1,13 +1,14 @@
 import type { PaneProps, PaneToggleProps } from '$lib/overlays'
 import type { D3ColorSchemeName, D3InterpolateName } from '$lib/colors'
 import type { D3SymbolName } from '$lib/labels'
-import type { Point2D, Point3D, Vec2, Vec3 } from '$lib/math'
+import type { Point2D, Vec2 } from '$lib/math'
 import type { ComponentProps, Snippet } from 'svelte'
 import type { HTMLAttributes } from 'svelte/elements'
 import type { TweenOptions } from 'svelte/motion'
 import type { Sides } from '$lib/plot/core/layout'
 import type PlotLegend from '$lib/plot/core/components/PlotLegend.svelte'
 import type { TicksOption } from '$lib/plot/core/scales'
+import type { FillGradient } from '$lib/plot/core/types/fills'
 
 export type { TweenOptions } from 'svelte/motion'
 
@@ -731,413 +732,8 @@ export const DEFAULT_SERIES_SYMBOLS = [
   `Wye`,
 ] as const satisfies readonly D3SymbolName[]
 
-// 3D point extending base Point with z coordinate (prefixed to avoid conflict with convex-hull)
-export interface ScatterPoint3D<Metadata = Record<string, unknown>> extends Point<Metadata> {
-  z: number
-}
-
-// 3D data series extending DataSeries with z array
-// Omit filtered_data since it uses 2D InternalPoint type, redeclare with 3D type
-export interface DataSeries3D<Metadata = Record<string, unknown>> extends Omit<
-  DataSeries<Metadata>,
-  `y_axis` | `filtered_data`
-> {
-  z: readonly number[]
-  filtered_data?: InternalPoint3D<Metadata>[]
-}
-
-// Internal 3D point for processing within ScatterPlot3D
-export interface InternalPoint3D<
-  Metadata = Record<string, unknown>,
-> extends ScatterPoint3D<Metadata> {
-  series_idx: number
-  point_idx: number
-  color_value?: number | null
-  size_value?: number | null
-  point_style?: PointStyle
-}
-
-// Surface types for 3D visualization
-export type SurfaceType = `grid` | `parametric` | `triangulated`
-
-// Configuration for 3D surfaces
-export interface Surface3DConfig {
-  id?: string | number
-  type: SurfaceType
-  // For grid surfaces: regular grid with z values
-  x_range?: Vec2
-  y_range?: Vec2
-  resolution?: number | Vec2 // grid resolution (x, y)
-  z_fn?: (x: number, y: number) => number
-  // For parametric surfaces: u,v parameterization
-  u_range?: Vec2
-  v_range?: Vec2
-  parametric_fn?: (u: number, v: number) => Point3D
-  // For triangulated surfaces: explicit geometry (only x,y,z needed, not scatter-specific fields)
-  points?: Point3D[]
-  triangles?: Vec3[] // indices into points array
-  // Appearance
-  color?: string
-  color_fn?: (x: number, y: number, z: number) => string
-  opacity?: number
-  wireframe?: boolean
-  wireframe_color?: string
-  wireframe_width?: number
-  visible?: boolean
-  // Double-sided rendering
-  double_sided?: boolean
-}
-
-// Extended axis config for 3D (same as 2D but can add 3D-specific options)
-export interface AxisConfig3D extends AxisConfig {
-  // 3D-specific axis options can be added here
-  show_plane?: boolean // Show grid plane for this axis
-  plane_opacity?: number
-}
-
-// Display config extended for 3D
-export interface DisplayConfig3D extends DisplayConfig {
-  z_grid?: boolean
-  z_zero_line?: boolean
-  show_axes?: boolean
-  show_axis_labels?: boolean
-  show_bounding_box?: boolean
-  show_grid?: boolean
-  // Projection settings - render point shadows on background planes
-  // Coordinate mapping: user X→Three.js X, user Y→Three.js Z, user Z→Three.js Y
-  projections?: {
-    xy?: boolean // Project onto XY plane (floor/ceiling) - fixes user Z
-    xz?: boolean // Project onto XZ plane (back wall) - fixes user Y
-    yz?: boolean // Project onto YZ plane (side wall) - fixes user X
-  }
-  projection_opacity?: number // 0-1, default 0.3
-  projection_scale?: number // Relative to point size, default 0.5
-}
-
-// 3D scatter handler props
-export interface Scatter3DHandlerProps<Metadata = Record<string, unknown>> extends Omit<
-  HandlerProps<Metadata>,
-  `x_axis` | `x2_axis` | `y_axis` | `y2_axis`
-> {
-  z: number
-  x_axis: AxisConfig3D
-  y_axis: AxisConfig3D
-  z_axis: AxisConfig3D
-  x_formatted: string
-  y_formatted: string
-  z_formatted: string
-  color_value?: number | null
-}
-
-export type Scatter3DHandlerEvent<Metadata = Record<string, unknown>> =
-  Scatter3DHandlerProps<Metadata> & {
-    event?: MouseEvent
-    point: InternalPoint3D<Metadata>
-  }
-
-// Camera projection types for 3D
-export type CameraProjection3D = `perspective` | `orthographic`
-
-// 3D plot config extending base
-export interface PlotConfig3D {
-  x_axis?: AxisConfig3D
-  y_axis?: AxisConfig3D
-  z_axis?: AxisConfig3D
-  display?: DisplayConfig3D
-}
-
-// 3D style overrides
-export interface StyleOverrides3D extends StyleOverrides {
-  point?: StyleOverrides[`point`] & {
-    sphere_segments?: number // Level of detail for sphere geometry
-  }
-}
-
-// Controls config for 3D
-export interface ControlsConfig3D extends ControlsConfig {
-  show_camera_controls?: boolean
-  show_surface_controls?: boolean
-}
-
-// FillBoundary - references to data sources for fill regions
-// Can reference series by index, by id, or specify constant/axis/function/data values
-export type FillBoundary =
-  | { type: `series`; series_idx: number }
-  | { type: `series`; series_id: string | number }
-  | { type: `constant`; value: number }
-  | { type: `axis`; axis: `x` | `x2` | `y` | `y2`; value?: number }
-  | { type: `function`; fn: (coord: number) => number }
-  // x is optional; when omitted, values align to the companion boundary's x positions
-  | { type: `data`; values: readonly number[]; x?: readonly number[] }
-  | number // Shorthand for constant value
-
-// Styling types for fill regions
-
-// Gradient configuration for fill regions
-export interface FillGradient {
-  type: `linear` | `radial`
-  // For linear gradients: angle in degrees (0 = horizontal left-to-right, 90 = vertical top-to-bottom)
-  angle?: number
-  // For radial gradients: center position (0-1 normalized)
-  center?: { x: number; y: number }
-  // Color stops: array of [offset (0-1), color] pairs
-  stops: readonly [number, string][]
-}
-
-// Edge/stroke styling for fill region boundaries
-export interface FillEdgeStyle {
-  color?: string
-  width?: number
-  dash?: string
-  opacity?: number
-}
-
-// Hover state styling for fill regions
-export interface FillHoverStyle {
-  fill?: string
-  fill_opacity?: number
-  edge?: FillEdgeStyle
-  cursor?: string
-  scale?: number // Scale factor for hover effect
-  stroke?: string // Outline color drawn around the region on hover (default: theme-aware contrast)
-  stroke_width?: number // Outline width in px on hover (default: 1.5)
-}
-
-// Event type for fill region interactions
-export interface FillHandlerEvent {
-  event: MouseEvent | KeyboardEvent
-  region_idx: number
-  region_id?: string | number
-  x: number // Data x-coordinate
-  y: number // Data y-coordinate
-  px: number // Pixel x-coordinate
-  py: number // Pixel y-coordinate
-  label?: string
-  metadata?: Record<string, unknown>
-}
-
-// z-index positions for rendering order (used by fill regions and reference lines)
-export type LayerZIndex = `below-grid` | `below-lines` | `below-points` | `above-all`
-
-// Curve types supported for fill path generation (matching d3-shape curves)
-export const FILL_CURVE_TYPES = [
-  `linear`,
-  `monotoneX`,
-  `monotoneY`,
-  `step`,
-  `stepBefore`,
-  `stepAfter`,
-  `basis`,
-  `cardinal`,
-  `catmullRom`,
-  `natural`,
-] as const
-export type FillCurveType = (typeof FILL_CURVE_TYPES)[number]
-
-// Main configuration for a fill region
-export interface FillRegion {
-  // Identity
-  id?: string | number
-  label?: string
-
-  // Boundaries - define the top and bottom of the fill area
-  upper: FillBoundary
-  lower: FillBoundary
-
-  // Optional range constraints
-  x_range?: [number | null, number | null]
-  y_range?: [number | null, number | null]
-
-  // Conditional fill: only fill where condition is true
-  // Example: where: (x, y_upper, y_lower) => y_upper > y_lower
-  where?: (x: number, y_upper: number, y_lower: number) => boolean
-
-  // Styling
-  fill?: string | FillGradient
-  fill_opacity?: number
-  edge_upper?: FillEdgeStyle
-  edge_lower?: FillEdgeStyle
-  curve?: FillCurveType
-  step_position?: number // For step curves: 0 = step, 0.5 = stepMiddle, 1 = stepEnd
-
-  // Rendering
-  z_index?: LayerZIndex
-  visible?: boolean
-
-  // Interactions
-  hover_style?: FillHoverStyle
-  on_click?: (event: FillHandlerEvent) => void
-  on_hover?: (event: FillHandlerEvent | null) => void
-
-  // Legend
-  show_in_legend?: boolean
-  legend_group?: string
-
-  // Metadata for user data
-  metadata?: Record<string, unknown>
-}
-
-// Convenience type for error bands (symmetric or asymmetric around a series)
-export interface ErrorBand {
-  // Reference to the central series
-  series:
-    | { type: `series`; series_idx: number }
-    | {
-        type: `series`
-        series_id: string | number
-      }
-
-  // Error values - can be symmetric (single value/array) or asymmetric (upper/lower)
-  error:
-    | number // Symmetric constant error
-    | readonly number[] // Symmetric per-point error
-    | { upper: number | readonly number[]; lower: number | readonly number[] } // Asymmetric
-
-  // Styling (defaults applied if not specified)
-  fill?: string // Defaults to series color with reduced opacity
-  fill_opacity?: number // Defaults to 0.3
-  edge_style?: FillEdgeStyle
-
-  // Identity
-  id?: string | number
-  label?: string
-  show_in_legend?: boolean
-}
-
-// Reference line styling
-export interface RefLineStyle {
-  color?: string // default: 'currentColor'
-  width?: number // default: 1
-  dash?: string // SVG stroke-dasharray, default: none (solid)
-  opacity?: number // default: 1
-}
-
-// Annotation/label for reference lines
-export interface RefLineAnnotation {
-  text: string
-  position?: `start` | `center` | `end` // position along the line
-  side?: `above` | `below` | `left` | `right` // which side of line
-  offset?: { x?: number; y?: number }
-  gap?: number // pixels between line and annotation text, default: 8
-  edge_padding?: number // pixels inward from plot edge at start/end, default: 4
-  font_size?: string
-  font_family?: string
-  color?: string
-  background?: string
-  padding?: number
-  rotate?: boolean // rotate text to follow line angle
-}
-
-// Event type for reference line interactions
-export interface RefLineEvent {
-  event: MouseEvent | KeyboardEvent | FocusEvent
-  line_idx: number
-  line_id?: string | number
-  type: RefLine[`type`]
-  label?: string
-  metadata?: Record<string, unknown>
-}
-
-// Base properties shared by all reference line types
-export interface RefLineBase {
-  id?: string | number
-  x_span?: [number | null, number | null]
-  y_span?: [number | null, number | null]
-  coord_mode?: `data` | `relative` // default: 'data'
-  x_axis?: `x1` | `x2` // for vertical lines with dual x-axes
-  y_axis?: `y1` | `y2` // for horizontal lines with dual y-axes
-  style?: RefLineStyle
-  annotation?: RefLineAnnotation
-  z_index?: LayerZIndex
-  visible?: boolean
-  hover_style?: RefLineStyle
-  on_click?: (event: RefLineEvent) => void
-  on_hover?: (event: RefLineEvent | null) => void
-  show_in_legend?: boolean
-  label?: string
-  legend_group?: string
-  metadata?: Record<string, unknown>
-}
-
-// Reference line value type - supports numbers, Dates, and ISO date strings
-export type RefLineValue = number | Date | string
-
-// Flat discriminated union - type determines required fields
-export type RefLine = RefLineBase &
-  (
-    | { type: `horizontal`; y: RefLineValue }
-    | { type: `vertical`; x: RefLineValue }
-    | { type: `diagonal`; slope: number; intercept: number }
-    | {
-        type: `segment`
-        p1: [RefLineValue, RefLineValue]
-        p2: [RefLineValue, RefLineValue]
-      }
-    | { type: `line`; p1: [RefLineValue, RefLineValue]; p2: [RefLineValue, RefLineValue] }
-  )
-
-// Default style values for reference lines
-export const REF_LINE_STYLE_DEFAULTS: Required<RefLineStyle> = {
-  color: `currentColor`,
-  width: 1,
-  dash: ``,
-  opacity: 1,
-} as const
-
-type Ref3DBase = Omit<
-  RefLineBase,
-  | `coord_mode`
-  | `x_axis`
-  | `y_axis`
-  | `style`
-  | `hover_style`
-  | `annotation`
-  | `on_click`
-  | `on_hover`
-> & {
-  z_span?: [number | null, number | null]
-}
-
-// Base properties shared by all 3D reference line types
-// Aligned with RefLineBase for future feature parity (interactions, annotations, etc.)
-export interface RefLine3DBase extends Ref3DBase {
-  style?: RefLineStyle
-  hover_style?: RefLineStyle
-  annotation?: RefLineAnnotation
-  on_click?: (event: { line_idx: number; line_id?: string | number }) => void
-  on_hover?: (event: { line_idx: number; line_id?: string | number } | null) => void
-}
-
-// 3D reference line - discriminated union
-export type RefLine3D = RefLine3DBase &
-  (
-    | { type: `x-axis`; y: number; z: number } // line parallel to x-axis
-    | { type: `y-axis`; x: number; z: number } // line parallel to y-axis
-    | { type: `z-axis`; x: number; y: number } // line parallel to z-axis
-    | { type: `segment`; p1: Vec3; p2: Vec3 }
-    | { type: `line`; p1: Vec3; p2: Vec3 }
-  )
-
-// 3D reference plane styling
-export interface RefPlaneStyle {
-  color?: string
-  opacity?: number
-  wireframe?: boolean
-  wireframe_color?: string
-  double_sided?: boolean
-}
-
-// Base properties shared by all 3D reference plane types
-export interface RefPlaneBase extends Ref3DBase {
-  style?: RefPlaneStyle
-}
-
-// 3D reference plane - discriminated union
-export type RefPlane = RefPlaneBase &
-  (
-    | { type: `xy`; z: number } // horizontal plane at z
-    | { type: `xz`; y: number } // vertical plane at y
-    | { type: `yz`; x: number } // vertical plane at x
-    | { type: `normal`; normal: Vec3; point: Vec3 }
-    | { type: `points`; p1: Vec3; p2: Vec3; p3: Vec3 } // plane through 3 points
-  )
+// Sub-domain types live in ./types/* and are re-exported so existing
+// `$lib/plot/core/types` import paths keep working.
+export type * from '$lib/plot/core/types/plot-3d'
+export * from '$lib/plot/core/types/fills'
+export * from '$lib/plot/core/types/reference-lines'
