+++ src/lib/plot/histogram/index.ts
@@ -1,2 +1,3 @@
 export { default as Histogram } from './Histogram.svelte'
 export { default as HistogramControls } from './HistogramControls.svelte'
+export * from './histogram'
